Refactor AMP implementation
